06_03

+++ src/com/kh/practice1/func/VariablePractice2.java
@@ -10,7 +10,7 @@ public class VariablePractice2 {
 		System.out.println("´õÇÏ±â °á°ú:" +(num1+num2));
 		System.out.println("»©±â °á°ú:" +(num1-num2));
 		System.out.println("°öÇÏ±â °á°ú:" +(num1*num2));
-		System.out.println("³ª´°¼À °á°ú:" +(num1/num2));
+		System.out.println("³ª´°¼À °á°ú:" +((double)num1/num2));
 	}
 	
 }
